+++ README.md
@@ -1,2 +1,10 @@
 # exploradome_tangram
 Tangram form detection from live video stream
+
+## reste à faire
+
+- convertir distances en probabilités
+- optimisation du code
+- élaborer stratégie de test : quels sont les éléments à tester ? combien d'éléments on prend en test ? métriques ?
+- minimisation du code
+- écriture du readme
