+++ app.py
@@ -3,6 +3,9 @@ import numpy as np
 import imutils
 import pandas as pd
 import moments as mm
+import os
+from distances import *
+from processing import *
 
 # main application
