+++ moments.py
@@ -3,74 +3,8 @@ import numpy as np
 import imutils
 import pandas as pd
 import os
-import time
-
-def get_files():
-    """
-    get images in jpg format in tangrams folder
-    """
-    images = {}
-    dirname = os.getcwd() + '/data/tangrams'
-    assert os.path.exists(dirname), "the directory doesn't exist"
-
-    for file in os.listdir(dirname):
-        filename, file_extension = os.path.splitext(file) # we just want the filename to save the path
-        if file.endswith((".jpg", ".JPG")):
-            images[filename] = os.path.join(dirname, file)
-    return images
-
-def resize(img, percent=20):
-    '''
-    this function takes a cv image as input and resizes it. The primary objective is to make the contouring less sensitive to between-tangram demarcation lines,
-    the secondary objective is to speed up processing.
-    '''
-    scale_percent = percent # percent of original size
-    width = int(img.shape[1] * scale_percent / 100)
-    height = int(img.shape[0] * scale_percent / 100)
-    dim = (width, height)
-    img = cv2.resize(img, dim, interpolation = cv2.INTER_AREA).copy()
-    return img
-
-def preprocess_img_test(img, sensitivity_to_light=50):
-    '''
-    this function takes a cv image as input, calls the resize function, crops the image to keep only the board, 
-    chooses the left or right half of the board, based on user input and eventually finds the largest dark shape
-    '''
-
-    # resize operations
-    img = resize(img, 20).copy()
-
-    # get the largest shape
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY) # binarize img
-    gray[gray>sensitivity_to_light] = 0 # turn background to black
-    blurred = cv2.GaussianBlur(gray, (5, 5), 0)
-    thresh = cv2.threshold(blurred, 0, 255, cv2.THRESH_BINARY)[1] # threshold ???
-    cnts = cv2.findContours(thresh.copy(), cv2.RETR_EXTERNAL,cv2.CHAIN_APPROX_SIMPLE)
-    cnts = imutils.grab_contours(cnts)
-    return cnts, img
-
-def preprocess_img(img, left_side=True, sensitivity_to_light=50):
-    '''
-    this function takes a cv image as input, calls the resize function, crops the image to keep only the board, 
-    chooses the left or right half of the board, based on user input and eventually finds the largest dark shape
-    '''
-
-    # resize operations
-    img = resize(img,20)[0:-50, 55:-100].copy() # resize img to 20% and crop to keep only board
-
-    if left_side:
-        img = img[0:int(img.shape[0]),0:int(img.shape[1]/2)] # keep only the left half of the board
-    else:
-        img = img[0:int(img.shape[0]),int(img.shape[1]/2):] # keep only the right half of the board
-    
-    # get the largest shape
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY) # binarize img
-    gray[gray>sensitivity_to_light] = 0 # turn background to black
-    blurred = cv2.GaussianBlur(gray, (5, 5), 0)
-    thresh = cv2.threshold(blurred, 0, 255, cv2.THRESH_BINARY)[1] # threshold ???
-    cnts = cv2.findContours(thresh.copy(), cv2.RETR_EXTERNAL,cv2.CHAIN_APPROX_SIMPLE)
-    cnts = imutils.grab_contours(cnts)
-    return cnts, img
+from distances import *
+from processing import *
 
 def find_moments(cnts, filename=None, hu_moment = True):
     '''
@@ -92,30 +26,6 @@ def find_moments(cnts, filename=None, hu_moment = True):
     if filename:
         Moms['target'] = filename
     return Moms
-
-def display_contour(cnts, img):
-    for c in cnts:
-        #compute the center of the contour
-        M = cv2.moments(c)
-        huM = cv2.HuMoments(M)
-        # draw the contour and center of the shape on the image
-        cv2.drawContours(img, [c], -1, (0, 255, 0), 2)
-    # show the image
-    cv2.imshow("Image", img)
-    cv2.waitKey(0)
-
-def dist_humoment1(hu1,hu2):
-    distance =  np.sum(abs(1/hu1-1/hu2))
-    return distance
-
-def dist_humoment2(hu1,hu2):
-    distance =  np.sum(abs(hu1-hu2))
-    # print(abs(hu1-hu2))
-    return distance
-
-def dist_humoment3(hu1,hu2):
-    distance =  np.sum(abs(hu1-hu2)/abs(hu2))
-    return distance
 
 def save_moments(images):
     """
@@ -140,26 +50,28 @@ def save_moments(images):
 
     return hu_moments, moments
     
-def test():
+def get_distance_img_test(image):
+    """
+    test the main algorithm on 1 image
+    """
     hu_moments = pd.read_csv('data/hu_moments.csv')
     target = hu_moments.iloc[:, -1]
 
     # get 1 img
     images = get_files()
 
-    # preprocess
-    img_cv = cv2.imread(images['bateau'])
+    # preprocessing the image
+    img_cv = cv2.imread(images[image])
     cnts, img_cv = preprocess_img_test(img_cv)
 
+    # compute distance
     HuMo = find_moments(cnts)
-    print(HuMo)
-
     proba = hu_moments.apply(lambda row : dist_humoment2(np.hstack(HuMo), row.values[:-1]), axis=1)
     proba_labelled = pd.concat([proba, target], axis=1)
     proba_labelled.columns = ['distance', 'target']
     print(proba_labelled.sort_values(by=["distance"]))
 
-def read_video():
+def read_video(video=False):
     """
     compare moments of each frame with the hu moments of our dataset images
     """
@@ -167,11 +79,17 @@ def read_video():
     hu_moments = pd.read_csv('data/hu_moments.csv')
     target = hu_moments.iloc[:, -1]
 
-    cap = cv2.VideoCapture(0)
+    if video:
+        cap = cv2.VideoCapture(video)
+    else : 
+        cap = cv2.VideoCapture(0)
 
-    while(True):
+    while(cap.isOpened()):
         # Capture frame-by-frame
         ret, image = cap.read()
+
+        if ret :
+            print(ret)
 
         # Our operations on the frame come here
         cnts, img = preprocess_img(image, left_side=False)
@@ -192,5 +110,4 @@ def read_video():
     cv2.destroyAllWindows()
 
 if __name__ == '__main__':
-    # test distance
-    test()
+    get_distance_img_test('cygne') # only for testing
